+++ example/serialise/Example.js
@@ -1,0 +1,7 @@
+export default class Example {
+  constructor(name, isExample) {
+    this.name = name
+    this.isExample = isExample
+    this.created = new Date(2019, 5)
+  }
+}